added validation chains and other access checks

+++ Services/ScheduleService.cs
@@ -5,18 +5,21 @@ using Microsoft.Extensions.Options;
 using MongoDB.Bson;
 using MongoDB.Driver;
 using MongoDotnetDemo.Models;
+using static System.Runtime.InteropServices.JavaScript.JSType;
 
 public class ScheduleService : IScheduleService
 {
     private readonly IMongoCollection<Schedule> _scheduleCollection;
     private readonly IOptions<DatabaseSettings> _dbSettings;
+    private readonly ITrainService _trainService;
 
-    public ScheduleService(IOptions<DatabaseSettings> dbSettings)
+    public ScheduleService(IOptions<DatabaseSettings> dbSettings , ITrainService trainService)
     {
         _dbSettings = dbSettings;
         var mongoClient = new MongoClient(dbSettings.Value.ConnectionString);
         var mongoDatabase = mongoClient.GetDatabase(dbSettings.Value.DatabaseName);
         _scheduleCollection = mongoDatabase.GetCollection<Schedule>(dbSettings.Value.SchedulesCollectionName);
+        _trainService = trainService;
     }
 
     public async Task<IEnumerable<Schedule>> GetAllAsync()
@@ -31,13 +34,21 @@ public class ScheduleService : IScheduleService
         return schedule;
     }
 
-    public async Task CreateAsync(Schedule schedule)
+    public async Task<string> CreateAsync(Schedule schedule)
     {
         await _scheduleCollection.InsertOneAsync(schedule);
+        return "Schedule created successfully";
     }
 
-    public async Task UpdateAsync(string id, Schedule schedule)
+    public async Task<string> UpdateAsync(string id, Schedule schedule)
     {
+        var existingSchedule = await _scheduleCollection.Find(t => t.Id == id).FirstOrDefaultAsync();
+        if (existingSchedule == null)
+            return "Schedule not found";
+
+        if (!await IsScheduleNotInUse(id))
+            return "Cannot update a schedule in use";
+        
         var filter = Builders<Schedule>.Filter.Eq(s => s.Id, id);
         var update = Builders<Schedule>.Update
             .Set(s => s.DepartureStation, schedule.DepartureStation)
@@ -47,10 +58,36 @@ public class ScheduleService : IScheduleService
             .Set(s => s.OperatingDays, schedule.OperatingDays);
 
         await _scheduleCollection.UpdateOneAsync(filter, update);
+        return "Schedule updated successfully";
     }
 
-    public async Task DeleteAsync(string id)
+    public async Task<string> DeleteAsync(string id)
     {
-        await _scheduleCollection.DeleteOneAsync(s => s.Id == id);
+        var existingSchedule = await _scheduleCollection.Find(t => t.Id == id).FirstOrDefaultAsync();
+        if (existingSchedule == null)
+            return "Schedule not found";
+
+        if (await IsScheduleNotInUse(id))
+        {
+            await _scheduleCollection.DeleteOneAsync(s => s.Id == id);
+            return "Schedule deleted successfully";
+        }
+        else
+        {
+            return "Cannot delete a schedule in use";
+        }
+    }
+
+    public async Task<bool> IsScheduleNotInUse(string id)
+    {
+        var trains = await _trainService.GetAllAsync();
+        bool isAssigned = true;
+
+        foreach (var train in trains)
+        {
+            if (train.Schedule == id)
+                isAssigned = false;
+        }
+        return isAssigned;
     }
 }
